Add PlayerManager and move player functions
