+++ src/Server.php
@@ -37,12 +37,7 @@ use pocketmine\crafting\CraftingManagerFromDataHelper;
 use pocketmine\crash\CrashDump;
 use pocketmine\crash\CrashDumpRenderer;
 use pocketmine\data\java\GameModeIdMap;
-use pocketmine\entity\EntityDataHelper;
-use pocketmine\entity\Location;
 use pocketmine\event\HandlerListManager;
-use pocketmine\event\player\PlayerCreationEvent;
-use pocketmine\event\player\PlayerDataSaveEvent;
-use pocketmine\event\player\PlayerLoginEvent;
 use pocketmine\event\server\CommandEvent;
 use pocketmine\event\server\DataPacketSendEvent;
 use pocketmine\event\server\QueryRegenerateEvent;
@@ -50,10 +45,6 @@ use pocketmine\lang\KnownTranslationFactory;
 use pocketmine\lang\Language;
 use pocketmine\lang\LanguageNotFoundException;
 use pocketmine\lang\Translatable;
-use pocketmine\nbt\BigEndianNbtSerializer;
-use pocketmine\nbt\NbtDataException;
-use pocketmine\nbt\tag\CompoundTag;
-use pocketmine\nbt\TreeRoot;
 use pocketmine\network\mcpe\compression\CompressBatchPromise;
 use pocketmine\network\mcpe\compression\CompressBatchTask;
 use pocketmine\network\mcpe\compression\Compressor;
@@ -74,9 +65,8 @@ use pocketmine\network\upnp\UPnPNetworkInterface;
 use pocketmine\permission\BanList;
 use pocketmine\permission\DefaultPermissions;
 use pocketmine\player\GameMode;
-use pocketmine\player\OfflinePlayer;
 use pocketmine\player\Player;
-use pocketmine\player\PlayerInfo;
+use pocketmine\player\PlayerManager;
 use pocketmine\plugin\PharPluginLoader;
 use pocketmine\plugin\Plugin;
 use pocketmine\plugin\PluginEnableOrder;
@@ -84,8 +74,6 @@ use pocketmine\plugin\PluginGraylist;
 use pocketmine\plugin\PluginManager;
 use pocketmine\plugin\PluginOwned;
 use pocketmine\plugin\ScriptPluginLoader;
-use pocketmine\promise\Promise;
-use pocketmine\promise\PromiseResolver;
 use pocketmine\resourcepacks\ResourcePackManager;
 use pocketmine\scheduler\AsyncPool;
 use pocketmine\snooze\SleeperHandler;
@@ -94,7 +82,6 @@ use pocketmine\stats\SendUsageTask;
 use pocketmine\timings\Timings;
 use pocketmine\timings\TimingsHandler;
 use pocketmine\updater\UpdateChecker;
-use pocketmine\utils\AssumptionFailedError;
 use pocketmine\utils\Config;
 use pocketmine\utils\Filesystem;
 use pocketmine\utils\Internet;
@@ -106,7 +93,6 @@ use pocketmine\utils\SignalHandler;
 use pocketmine\utils\Terminal;
 use pocketmine\utils\TextFormat;
 use pocketmine\utils\Utils;
-use pocketmine\world\format\Chunk;
 use pocketmine\world\format\io\WorldProviderManager;
 use pocketmine\world\format\io\WritableWorldProviderManagerEntry;
 use pocketmine\world\generator\Generator;
@@ -151,22 +137,15 @@ use function spl_object_id;
 use function sprintf;
 use function str_repeat;
 use function str_replace;
-use function stripos;
-use function strlen;
 use function strrpos;
-use function strtolower;
 use function strval;
 use function time;
 use function touch;
 use function trim;
 use function yaml_parse;
-use function zlib_decode;
-use function zlib_encode;
 use const DIRECTORY_SEPARATOR;
 use const PHP_EOL;
-use const PHP_INT_MAX;
 use const PTHREADS_INHERIT_NONE;
-use const ZLIB_ENCODING_GZIP;
 
 /**
  * The class that manages everything
@@ -181,14 +160,6 @@ class Server{
 	private static ?Server $instance = null;
 
 	private SleeperHandler $tickSleeper;
-
-	private BanList $banByName;
-
-	private BanList $banByIP;
-
-	private Config $operators;
-
-	private Config $whitelist;
 
 	private bool $isRunning = true;
 
@@ -247,18 +218,9 @@ class Server{
 	private string $dataPath;
 	private string $pluginPath;
 
-	/**
-	 * @var string[]
-	 * @phpstan-var array<string, string>
-	 */
-	private array $uniquePlayers = [];
-
 	private QueryInfo $queryInfo;
 
 	private ServerConfigGroup $configGroup;
-
-	/** @var Player[] */
-	private array $playerList = [];
 
 	private SignalHandler $signalHandler;
 
@@ -455,211 +417,6 @@ class Server{
 		return $this->commandMap;
 	}
 
-	/**
-	 * @return Player[]
-	 */
-	public function getOnlinePlayers() : array{
-		return $this->playerList;
-	}
-
-	public function shouldSavePlayerData() : bool{
-		return $this->configGroup->getPropertyBool("player.save-player-data", true);
-	}
-
-	/**
-	 * @return OfflinePlayer|Player
-	 */
-	public function getOfflinePlayer(string $name){
-		$name = strtolower($name);
-		$result = $this->getPlayerExact($name);
-
-		if($result === null){
-			$result = new OfflinePlayer($name, $this->getOfflinePlayerData($name));
-		}
-
-		return $result;
-	}
-
-	private function getPlayerDataPath(string $username) : string{
-		return Path::join($this->getDataPath(), 'players', strtolower($username) . '.dat');
-	}
-
-	/**
-	 * Returns whether the server has stored any saved data for this player.
-	 */
-	public function hasOfflinePlayerData(string $name) : bool{
-		return file_exists($this->getPlayerDataPath($name));
-	}
-
-	private function handleCorruptedPlayerData(string $name) : void{
-		$path = $this->getPlayerDataPath($name);
-		rename($path, $path . '.bak');
-		$this->logger->error($this->getLanguage()->translate(KnownTranslationFactory::pocketmine_data_playerCorrupted($name)));
-	}
-
-	public function getOfflinePlayerData(string $name) : ?CompoundTag{
-		return Timings::$syncPlayerDataLoad->time(function() use ($name) : ?CompoundTag{
-			$name = strtolower($name);
-			$path = $this->getPlayerDataPath($name);
-
-			if(file_exists($path)){
-				$contents = @file_get_contents($path);
-				if($contents === false){
-					throw new \RuntimeException("Failed to read player data file \"$path\" (permission denied?)");
-				}
-				$decompressed = @zlib_decode($contents);
-				if($decompressed === false){
-					$this->logger->debug("Failed to decompress raw player data for \"$name\"");
-					$this->handleCorruptedPlayerData($name);
-					return null;
-				}
-
-				try{
-					return (new BigEndianNbtSerializer())->read($decompressed)->mustGetCompoundTag();
-				}catch(NbtDataException $e){ //corrupt data
-					$this->logger->debug("Failed to decode NBT data for \"$name\": " . $e->getMessage());
-					$this->handleCorruptedPlayerData($name);
-					return null;
-				}
-			}
-			return null;
-		});
-	}
-
-	public function saveOfflinePlayerData(string $name, CompoundTag $nbtTag) : void{
-		$ev = new PlayerDataSaveEvent($nbtTag, $name, $this->getPlayerExact($name));
-		if(!$this->shouldSavePlayerData()){
-			$ev->cancel();
-		}
-
-		$ev->call();
-
-		if(!$ev->isCancelled()){
-			Timings::$syncPlayerDataSave->time(function() use ($name, $ev) : void{
-				$nbt = new BigEndianNbtSerializer();
-				$contents = zlib_encode($nbt->write(new TreeRoot($ev->getSaveData())), ZLIB_ENCODING_GZIP);
-				if($contents === false){
-					throw new AssumptionFailedError("zlib_encode() failed unexpectedly");
-				}
-				try{
-					Filesystem::safeFilePutContents($this->getPlayerDataPath($name), $contents);
-				}catch(\RuntimeException | \ErrorException $e){
-					$this->logger->critical($this->getLanguage()->translate(KnownTranslationFactory::pocketmine_data_saveError($name, $e->getMessage())));
-					$this->logger->logException($e);
-				}
-			});
-		}
-	}
-
-	/**
-	 * @phpstan-return Promise<Player>
-	 */
-	public function createPlayer(NetworkSession $session, PlayerInfo $playerInfo, bool $authenticated, ?CompoundTag $offlinePlayerData) : Promise{
-		$ev = new PlayerCreationEvent($session);
-		$ev->call();
-		$class = $ev->getPlayerClass();
-
-		if($offlinePlayerData !== null and ($world = $this->worldManager->getWorldByName($offlinePlayerData->getString("Level", ""))) !== null){
-			$playerPos = EntityDataHelper::parseLocation($offlinePlayerData, $world);
-			$spawn = $playerPos->asVector3();
-		}else{
-			$world = $this->worldManager->getDefaultWorld();
-			if($world === null){
-				throw new AssumptionFailedError("Default world should always be loaded");
-			}
-			$playerPos = null;
-			$spawn = $world->getSpawnLocation();
-		}
-		$playerPromiseResolver = new PromiseResolver();
-		$world->requestChunkPopulation($spawn->getFloorX() >> Chunk::COORD_BIT_SIZE, $spawn->getFloorZ() >> Chunk::COORD_BIT_SIZE, null)->onCompletion(
-			function() use ($playerPromiseResolver, $class, $session, $playerInfo, $authenticated, $world, $playerPos, $spawn, $offlinePlayerData) : void{
-				if(!$session->isConnected()){
-					$playerPromiseResolver->reject();
-					return;
-				}
-
-				/* Stick with the original spawn at the time of generation request, even if it changed since then.
-				 * This is because we know for sure that that chunk will be generated, but the one at the new location
-				 * might not be, and it would be much more complex to go back and redo the whole thing.
-				 *
-				 * TODO: this relies on the assumption that getSafeSpawn() will only alter the Y coordinate of the
-				 * provided position. If this assumption is broken, we'll start seeing crashes in here.
-				 */
-
-				/**
-				 * @see Player::__construct()
-				 * @var Player $player
-				 */
-				$player = new $class($this, $session, $playerInfo, $authenticated, $playerPos ?? Location::fromObject($world->getSafeSpawn($spawn), $world), $offlinePlayerData);
-				if(!$player->hasPlayedBefore()){
-					$player->onGround = true;  //TODO: this hack is needed for new players in-air ticks - they don't get detected as on-ground until they move
-				}
-				$playerPromiseResolver->resolve($player);
-			},
-			static function() use ($playerPromiseResolver, $session) : void{
-				if($session->isConnected()){
-					$session->disconnect("Spawn terrain generation failed");
-				}
-				$playerPromiseResolver->reject();
-			}
-		);
-		return $playerPromiseResolver->getPromise();
-	}
-
-	/**
-	 * Returns an online player whose name begins with or equals the given string (case insensitive).
-	 * The closest match will be returned, or null if there are no online matches.
-	 *
-	 * @see Server::getPlayerExact()
-	 */
-	public function getPlayerByPrefix(string $name) : ?Player{
-		$found = null;
-		$name = strtolower($name);
-		$delta = PHP_INT_MAX;
-		foreach($this->getOnlinePlayers() as $player){
-			if(stripos($player->getName(), $name) === 0){
-				$curDelta = strlen($player->getName()) - strlen($name);
-				if($curDelta < $delta){
-					$found = $player;
-					$delta = $curDelta;
-				}
-				if($curDelta === 0){
-					break;
-				}
-			}
-		}
-
-		return $found;
-	}
-
-	/**
-	 * Returns an online player with the given name (case insensitive), or null if not found.
-	 */
-	public function getPlayerExact(string $name) : ?Player{
-		$name = strtolower($name);
-		foreach($this->getOnlinePlayers() as $player){
-			if(strtolower($player->getName()) === $name){
-				return $player;
-			}
-		}
-
-		return null;
-	}
-
-	/**
-	 * Returns the player online with the specified raw UUID, or null if not found
-	 */
-	public function getPlayerByRawUUID(string $rawUUID) : ?Player{
-		return $this->playerList[$rawUUID] ?? null;
-	}
-
-	/**
-	 * Returns the player online with a UUID equivalent to the specified UuidInterface object, or null if not found
-	 */
-	public function getPlayerByUUID(UuidInterface $uuid) : ?Player{
-		return $this->getPlayerByRawUUID($uuid->getBytes());
-	}
-
 	public function getConfigGroup() : ServerConfigGroup{
 		return $this->configGroup;
 	}
@@ -674,64 +431,6 @@ class Server{
 		}else{
 			return null;
 		}
-	}
-
-	public function getNameBans() : BanList{
-		return $this->banByName;
-	}
-
-	public function getIPBans() : BanList{
-		return $this->banByIP;
-	}
-
-	public function addOp(string $name) : void{
-		$this->operators->set(strtolower($name), true);
-
-		if(($player = $this->getPlayerExact($name)) !== null){
-			$player->setBasePermission(DefaultPermissions::ROOT_OPERATOR, true);
-		}
-		$this->operators->save();
-	}
-
-	public function removeOp(string $name) : void{
-		$lowercaseName = strtolower($name);
-		foreach($this->operators->getAll() as $operatorName => $_){
-			$operatorName = (string) $operatorName;
-			if($lowercaseName === strtolower($operatorName)){
-				$this->operators->remove($operatorName);
-			}
-		}
-
-		if(($player = $this->getPlayerExact($name)) !== null){
-			$player->unsetBasePermission(DefaultPermissions::ROOT_OPERATOR);
-		}
-		$this->operators->save();
-	}
-
-	public function addWhitelist(string $name) : void{
-		$this->whitelist->set(strtolower($name), true);
-		$this->whitelist->save();
-	}
-
-	public function removeWhitelist(string $name) : void{
-		$this->whitelist->remove(strtolower($name));
-		$this->whitelist->save();
-	}
-
-	public function isWhitelisted(string $name) : bool{
-		return !$this->hasWhitelist() or $this->operators->exists($name, true) or $this->whitelist->exists($name, true);
-	}
-
-	public function isOp(string $name) : bool{
-		return $this->operators->exists($name, true);
-	}
-
-	public function getWhitelisted() : Config{
-		return $this->whitelist;
-	}
-
-	public function getOps() : Config{
-		return $this->operators;
 	}
 
 	/**
@@ -1184,7 +883,7 @@ class Server{
 			$this->network->registerRawPacketHandler(new QueryHandler($this));
 		}
 
-		foreach($this->getIPBans()->getEntries() as $entry){
+		foreach($this->getPlayerManager()->getIPBans()->getEntries() as $entry){
 			$this->network->blockAddress($entry->getName(), -1);
 		}
 
@@ -1658,55 +1357,25 @@ class Server{
 		}
 	}
 
-	public function addOnlinePlayer(Player $player) : bool{
-		$ev = new PlayerLoginEvent($player, "Plugin reason");
-		$ev->call();
-		if($ev->isCancelled() or !$player->isConnected()){
-			$player->disconnect($ev->getKickMessage());
-
-			return false;
-		}
-
-		$session = $player->getNetworkSession();
-		$position = $player->getPosition();
-		$this->logger->info($this->language->translate(KnownTranslationFactory::pocketmine_player_logIn(
-			TextFormat::AQUA . $player->getName() . TextFormat::WHITE,
-			$session->getIp(),
-			(string) $session->getPort(),
-			(string) $player->getId(),
-			$position->getWorld()->getDisplayName(),
-			(string) round($position->x, 4),
-			(string) round($position->y, 4),
-			(string) round($position->z, 4)
-		)));
-
-		foreach($this->playerList as $p){
-			$p->getNetworkSession()->onPlayerAdded($player);
-		}
-		$rawUUID = $player->getUniqueId()->getBytes();
-		$this->playerList[$rawUUID] = $player;
-
-		if($this->sendUsageTicker > 0){
-			$this->uniquePlayers[$rawUUID] = $rawUUID;
-		}
-
-		return true;
-	}
-
-	public function removeOnlinePlayer(Player $player) : void{
-		if(isset($this->playerList[$rawUUID = $player->getUniqueId()->getBytes()])){
-			unset($this->playerList[$rawUUID]);
-			foreach($this->playerList as $p){
-				$p->getNetworkSession()->onPlayerRemoved($player);
-			}
-		}
-	}
-
 	public function sendUsage(int $type = SendUsageTask::TYPE_STATUS) : void{
 		if($this->configGroup->getPropertyBool("anonymous-statistics.enabled", true)){
-			$this->asyncPool->submitTask(new SendUsageTask($this, $type, $this->uniquePlayers));
+			$this->asyncPool->submitTask(new SendUsageTask($this, $type, PlayerManager::getUniquePlayers()));
 		}
-		$this->uniquePlayers = [];
+		PlayerManager::$uniquePlayers = [];
+	}
+
+	/**
+	 * @return PlayerManager
+	 */
+	public function getPlayerManager() : PlayerManager{
+		return new PlayerManager($this);
+	}
+
+	/**
+	 * @return int
+	 */
+	public function getSendUsageTicker() : int{
+		return $this->sendUsageTicker;
 	}
 
 	public function getLanguage() : Language{
@@ -1731,7 +1400,7 @@ class Server{
 		$u = Process::getAdvancedMemoryUsage();
 		$usage = sprintf("%g/%g/%g MB @ %d threads", round(($u[0] / 1024) / 1024, 2), round(($u[1] / 1024) / 1024, 2), round(($u[2] / 1024) / 1024, 2), Process::getThreadCount());
 
-		$online = count($this->playerList);
+		$online = count($this->getPlayerManager()->getPlayerList());
 		$connecting = $this->network->getConnectionCount() - $online;
 		$bandwidthStats = $this->network->getBandwidthTracker();
