+++ src/command/defaults/ClearCommand.php
@@ -59,7 +59,7 @@ class ClearCommand extends VanillaCommand{
 
 		$target = null;
 		if(isset($args[0])){
-			$target = $sender->getServer()->getPlayerByPrefix($args[0]);
+			$target = $sender->getServer()->getPlayerManager()->getPlayerByPrefix($args[0]);
 			if($target === null){
 				$sender->sendMessage(KnownTranslationFactory::commands_generic_player_notFound()->prefix(TextFormat::RED));
 				return true;
