+++ src/command/defaults/EffectCommand.php
@@ -54,7 +54,7 @@ class EffectCommand extends VanillaCommand{
 			throw new InvalidCommandSyntaxException();
 		}
 
-		$player = $sender->getServer()->getPlayerByPrefix($args[0]);
+		$player = $sender->getServer()->getPlayerManager()->getPlayerByPrefix($args[0]);
 
 		if($player === null){
 			$sender->sendMessage(KnownTranslationFactory::commands_generic_player_notFound()->prefix(TextFormat::RED));
