+++ src/command/defaults/EnchantCommand.php
@@ -52,7 +52,7 @@ class EnchantCommand extends VanillaCommand{
 			throw new InvalidCommandSyntaxException();
 		}
 
-		$player = $sender->getServer()->getPlayerByPrefix($args[0]);
+		$player = $sender->getServer()->getPlayerManager()->getPlayerByPrefix($args[0]);
 
 		if($player === null){
 			$sender->sendMessage(KnownTranslationFactory::commands_generic_player_notFound()->prefix(TextFormat::RED));
