+++ src/command/defaults/KillCommand.php
@@ -60,7 +60,7 @@ class KillCommand extends VanillaCommand{
 				return true;
 			}
 
-			$player = $sender->getServer()->getPlayerByPrefix($args[0]);
+			$player = $sender->getServer()->getPlayerManager()->getPlayerByPrefix($args[0]);
 
 			if($player instanceof Player){
 				$player->attack(new EntityDamageEvent($player, EntityDamageEvent::CAUSE_SUICIDE, 1000));
