+++ src/command/defaults/TeleportCommand.php
@@ -50,7 +50,7 @@ class TeleportCommand extends VanillaCommand{
 	}
 
 	private function findPlayer(CommandSender $sender, string $playerName) : ?Player{
-		$subject = $sender->getServer()->getPlayerByPrefix($playerName);
+		$subject = $sender->getServer()->getPlayerManager()->getPlayerByPrefix($playerName);
 		if($subject === null){
 			$sender->sendMessage(TextFormat::RED . "Can't find player " . $playerName);
 			return null;
