+++ src/command/defaults/TellCommand.php
@@ -55,7 +55,7 @@ class TellCommand extends VanillaCommand{
 			throw new InvalidCommandSyntaxException();
 		}
 
-		$player = $sender->getServer()->getPlayerByPrefix(array_shift($args));
+		$player = $sender->getServer()->getPlayerManager()->getPlayerByPrefix(array_shift($args));
 
 		if($player === $sender){
 			$sender->sendMessage(KnownTranslationFactory::commands_message_sameTarget()->prefix(TextFormat::RED));
