+++ src/stats/SendUsageTask.php
@@ -125,7 +125,7 @@ class SendUsageTask extends AsyncTask{
 				//This anonymizes the user ids so they cannot be reversed to the original
 				$playerList = array_map('md5', $playerList);
 
-				$players = array_map(function(Player $p) : string{ return md5($p->getUniqueId()->getBytes()); }, $server->getOnlinePlayers());
+				$players = array_map(function(Player $p) : string{ return md5($p->getUniqueId()->getBytes()); }, $server->getPlayerManager()->getOnlinePlayers());
 
 				$data["players"] = [
 					"count" => count($players),
